Added a backend update profile

+++ 4-mern/e-com/backend/seeder/seeder.js
@@ -1,6 +1,7 @@
 import mongoose from "mongoose";
 import products from "./data.js";
 import Product from "../models/product.js"
+import User from "../models/user.js";
 
 
 const seedProducts = async () => {
@@ -9,14 +10,27 @@ const seedProducts = async () => {
 
         await mongoose.connect("mongodb+srv://javid:javid123@ecommerce.wutpy.mongodb.net/ecom");
 
+        // Fetch the first user (admin)
+
+        const firstUser = await User.findOne();
+
+        if(!firstUser){
+            console.log("No User Found");
+            process.exit(1);
+        }
+
+
         // Delete existing products
         await Product.deleteMany();
 
-        console.log("Products are deleted");
+        // Assing user ID to Products
+        const updateProducts = products.map((product) => ({
+            ...product,
+            user:firstUser._id
+        }))
 
-        // Insert a New Products
-        await Product.insertMany(products);
-        console.log("Products are added");
+        // Insert New Products 
+        await Product.insertMany(updateProducts);
 
         process.exit();
